Split lib test suites into integration targets (Fixes #307)
Move GitHub client, layout, and other public behavioral suites out of
the lib harness so Clippy's large_stack_arrays ceiling no longer blocks
new independent tests. Document the per-target registration budget.
