--- src/github/tests/mod.rs
@@ -1,5 +1,0 @@
-mod actions;
-mod auth_device;
-mod create_issue;
-mod issues;
-mod state_reason;
