+++ tests/agent_detection.rs
@@ -1,0 +1,4 @@
+//! Agent detection tests moved out of the lib target (issue #307).
+
+#[path = "agent_detection/detection_tests.rs"]
+mod detection_tests;
