+++ tests/identity.rs
@@ -1,0 +1,4 @@
+//! Process identity tests moved out of the lib target (issue #307).
+
+#[path = "identity/identity_tests.rs"]
+mod identity_tests;
