+++ tests/layout.rs
@@ -1,0 +1,4 @@
+//! Layout unit tests moved out of the lib target (issue #307).
+
+#[path = "layout/layout_tests.rs"]
+mod layout_tests;
