+++ tests/layout/layout_tests.rs
@@ -6,7 +6,15 @@
 //! @requirement REQ-PR-006
 //! @requirement REQ-PR-009
 
-use super::*;
+use jefe::layout::*;
+
+fn effective_render_size_for_test(cols: u16, rows: u16, fullscreen: bool) -> (u16, u16) {
+    effective_render_size_for_windowed(cols, rows, !fullscreen)
+}
+
+fn compute_pty_layout_for_test(cols: u16, rows: u16, fullscreen: bool) -> PtyLayout {
+    compute_pty_layout_for_windowed(cols, rows, !fullscreen)
+}
 
 #[test]
 fn column_width_constants_hold_expected_values() {
@@ -20,22 +28,22 @@ fn column_width_constants_hold_expected_values() {
 
 #[test]
 fn effective_render_size_fullscreen_passthrough() {
-    assert_eq!(effective_render_size_inner(120, 40, true), (120, 40));
-    assert_eq!(effective_render_size_inner(80, 24, true), (80, 24));
+    assert_eq!(effective_render_size_for_test(120, 40, true), (120, 40));
+    assert_eq!(effective_render_size_for_test(80, 24, true), (80, 24));
 }
 
 #[test]
 fn effective_render_size_windowed_subtraction() {
-    assert_eq!(effective_render_size_inner(120, 40, false), (118, 38));
-    assert_eq!(effective_render_size_inner(2, 2, false), (1, 1));
-    assert_eq!(effective_render_size_inner(1, 1, false), (1, 1));
+    assert_eq!(effective_render_size_for_test(120, 40, false), (118, 38));
+    assert_eq!(effective_render_size_for_test(2, 2, false), (1, 1));
+    assert_eq!(effective_render_size_for_test(1, 1, false), (1, 1));
 }
 
 #[test]
 fn fullscreen_and_windowed_terminals_can_project_the_same_render_size() {
     assert_eq!(
-        effective_render_size_inner(100, 25, true),
-        effective_render_size_inner(102, 27, false)
+        effective_render_size_for_test(100, 25, true),
+        effective_render_size_for_test(102, 27, false)
     );
 }
 
@@ -82,7 +90,7 @@ fn terminal_manager_pty_layout_matches_lower_workspace_pane() {
 
 #[test]
 fn compute_pty_layout_pane_origin() {
-    let layout = compute_pty_layout_inner(120, 40, true);
+    let layout = compute_pty_layout_for_test(120, 40, true);
     assert_eq!(layout.pane_col0, LEFT_COL_WIDTH + 1);
 }
 
@@ -106,7 +114,7 @@ fn dashboard_middle_row_heights_degrade_gracefully_when_extremely_small() {
 fn compute_pty_layout_dimensions_always_at_least_two() {
     for fullscreen in [true, false] {
         for (cols, rows) in [(120, 40), (10, 10), (0, 0), (60, 20)] {
-            let layout = compute_pty_layout_inner(cols, rows, fullscreen);
+            let layout = compute_pty_layout_for_test(cols, rows, fullscreen);
             assert!(
                 layout.pty_rows >= 2,
                 "pty_rows < 2 for ({cols}, {rows}, fullscreen={fullscreen})"
@@ -122,7 +130,7 @@ fn compute_pty_layout_dimensions_always_at_least_two() {
 #[test]
 fn agent_rows_rounding_half_up_fullscreen() {
     // 40 rows - 2 bars = 38 content rows. 25% = 9.5 → rounds to 10.
-    let layout = compute_pty_layout_inner(120, 40, true);
+    let layout = compute_pty_layout_for_test(120, 40, true);
     // pane_row0 = 1 (status bar) + agent_rows(10) + 2 (chrome top border + header)
     assert_eq!(layout.pane_row0, 1 + 10 + 2);
 }
@@ -130,14 +138,14 @@ fn agent_rows_rounding_half_up_fullscreen() {
 #[test]
 fn agent_rows_rounding_half_up_windowed() {
     // Windowed: 40-2=38 render rows, 38-2=36 content rows. 25% = 9.0 → exactly 9.
-    let layout = compute_pty_layout_inner(120, 40, false);
+    let layout = compute_pty_layout_for_test(120, 40, false);
     assert_eq!(layout.pane_row0, 1 + 9 + 2);
 }
 
 #[test]
 fn compute_pty_layout_pane_row0_positive() {
     for fullscreen in [true, false] {
-        let layout = compute_pty_layout_inner(120, 40, fullscreen);
+        let layout = compute_pty_layout_for_test(120, 40, fullscreen);
         assert!(
             layout.pane_row0 > 0,
             "pane_row0 not positive for fullscreen={fullscreen}"
@@ -384,7 +392,7 @@ fn prop_pty_dimensions_invariants_hold_across_sizes() {
     for fullscreen in [true, false] {
         for &cols in &COL_SAMPLES {
             for &rows in &ROW_SAMPLES {
-                let layout = compute_pty_layout_inner(cols, rows, fullscreen);
+                let layout = compute_pty_layout_for_test(cols, rows, fullscreen);
                 assert!(
                     layout.pty_rows >= 2,
                     "pty_rows < 2 for ({cols}, {rows}, fs={fullscreen})"
@@ -396,7 +404,7 @@ fn prop_pty_dimensions_invariants_hold_across_sizes() {
             }
             // Dense row sweep: every value 0..=64, both fullscreen states.
             for rows in 0..=64u16 {
-                let layout = compute_pty_layout_inner(cols, rows, fullscreen);
+                let layout = compute_pty_layout_for_test(cols, rows, fullscreen);
                 assert!(
                     layout.pty_rows >= 2,
                     "pty_rows < 2 for (cols={cols}, rows={rows}, fs={fullscreen})"
@@ -415,7 +423,7 @@ fn prop_pane_origin_invariants() {
     for fullscreen in [true, false] {
         for &cols in &COL_SAMPLES {
             for &rows in &ROW_SAMPLES {
-                let layout = compute_pty_layout_inner(cols, rows, fullscreen);
+                let layout = compute_pty_layout_for_test(cols, rows, fullscreen);
                 assert_eq!(
                     layout.pane_col0,
                     LEFT_COL_WIDTH + 1,
@@ -441,10 +449,10 @@ fn prop_agent_rows_half_up_rounding() {
     for fullscreen in [true, false] {
         for term_rows in 0..=300u16 {
             let cols: u16 = 120; // wide enough that cols don't constrain rows
-            let (_, eff_rows) = effective_render_size_inner(cols, term_rows, fullscreen);
+            let (_, eff_rows) = effective_render_size_for_test(cols, term_rows, fullscreen);
             let content_rows = eff_rows.saturating_sub(OUTER_BARS_HEIGHT);
             let agent_rows = expected_agent_rows(content_rows);
-            let layout = compute_pty_layout_inner(cols, term_rows, fullscreen);
+            let layout = compute_pty_layout_for_test(cols, term_rows, fullscreen);
             // pane_row0 = 1 + agent_rows + 2
             let expected_pane_row0 = 1u16.saturating_add(agent_rows).saturating_add(2);
             assert_eq!(
@@ -670,7 +678,7 @@ const GOLDEN_CASES: &[(u16, u16, bool, PtyLayout)] = &[
 #[test]
 fn golden_pty_layout_representative_sizes() {
     for &(cols, rows, fullscreen, expected) in GOLDEN_CASES {
-        let actual = compute_pty_layout_inner(cols, rows, fullscreen);
+        let actual = compute_pty_layout_for_test(cols, rows, fullscreen);
         assert_eq!(
             actual, expected,
             "golden mismatch for ({cols}x{rows}, fullscreen={fullscreen})"
