+++ tests/markdown_html_strip.rs
@@ -1,0 +1,4 @@
+//! HTML strip state-machine tests moved out of the lib target (issue #307).
+
+#[path = "markdown_html_strip/strip_tests.rs"]
+mod strip_tests;
