+++ tests/pagination.rs
@@ -1,0 +1,4 @@
+//! Pagination contract tests moved out of the lib target (issue #307).
+
+#[path = "pagination/page_token_tests.rs"]
+mod page_token_tests;
